4 December 2020, Norm'ing: split redir_basic_states.c into two files. One with all the quote states (redir_quote_states.c) and one with the basic and backslash states (redir_basis_and_escape_states.c).
